+++ packages/expo-dev-launcher/plugin/src/withDevLauncher.ts
@@ -218,7 +218,16 @@ const withDevLauncherPodfile: ConfigPlugin = config => {
     async config => {
       await editPodfile(config, podfile => {
         podfile = podfile.replace("platform :ios, '10.0'", "platform :ios, '11.0'");
-        podfile = addLines(podfile, 'use_react_native', 0, [`  ${DEV_LAUNCHER_POD_IMPORT}`]);
+        // Match both variations of Ruby config:
+        // unknown: pod 'expo-dev-menu', path: '../node_modules/expo-dev-menu', :configurations => :debug
+        // Rubocop: pod 'expo-dev-menu', path: '../node_modules/expo-dev-menu', configurations: :debug
+        if (
+          !podfile.match(
+            /pod ['"]expo-dev-menu['"],\s?path: ['"]\.\.\/node_modules\/expo-dev-menu['"],\s?:?configurations:?\s(?:=>\s)?:debug/
+          )
+        ) {
+          podfile = addLines(podfile, 'use_react_native', 0, [`  ${DEV_LAUNCHER_POD_IMPORT}`]);
+        }
         return podfile;
       });
       return config;